Move adjust_spots.html into its own directory. Add create_spots.py.
The python script create_spots.py can be used to create example
input JSON files for adjust_spots.html. In the JSON format there
is a now a "radius" field for each spot. The width of a connected
line is now calculated from the radiuses of the two neighbouring
circles.

The spots.json file was updated. It now contains the "radius" field.